+++ Realm/RLMArray.mm
@@ -543,6 +543,10 @@ static bool canAggregate(RLMPropertyType type, bool allowDate) {
 // http://www.openradar.me/radar?id=6135653276319744
 #pragma clang diagnostic ignored "-Wmismatched-parameter-types"
 - (RLMNotificationToken *)addNotificationBlock:(void (^)(RLMArray *, RLMCollectionChange *, NSError *))block {
+    return [self addNotificationBlock:block receiveOnQueue:nil];
+}
+- (RLMNotificationToken *)addNotificationBlock:(void (^)(RLMArray *, RLMCollectionChange *, NSError *))block
+                                receiveOnQueue:(nullable dispatch_queue_t)queue {
     @throw RLMException(@"This method may only be called on RLMArray instances retrieved from an RLMRealm");
 }