+++ RealmSwift/LinkingObjects.swift
@@ -258,9 +258,10 @@ public struct LinkingObjects<Element: Object> {
      not perform a write transaction on the same thread or explicitly call `realm.refresh()`, accessing it will never
      perform blocking work.
 
-     Notifications are delivered via the standard run loop, and so can't be delivered while the run loop is blocked by
-     other activity. When notifications can't be delivered instantly, multiple notifications may be coalesced into a
-     single notification. This can include the notification with the initial collection.
+     If no queue is given, notifications are delivered via the standard run loop, and so can't be delivered while the
+     run loop is blocked by other activity. If a queue is given, notifications are delivered to that queue instead. When
+     notifications can't be delivered instantly, multiple notifications may be coalesced into a single notification.
+     This can include the notification with the initial collection.
 
      For example, the following code performs a write transaction immediately after adding the notification block, so
      there is no opportunity for the initial notification to be delivered first. As a result, the initial notification
@@ -295,13 +296,14 @@ public struct LinkingObjects<Element: Object> {
 
      - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
 
+     - parameter queue: The serial dispatch queue to receive notification on. If
+                        `nil`, notifications are delivered to the current thread.
      - parameter block: The block to be called whenever a change occurs.
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
-    public func observe(_ block: @escaping (RealmCollectionChange<LinkingObjects>) -> Void) -> NotificationToken {
-        return rlmResults.addNotificationBlock { _, change, error in
-            block(RealmCollectionChange.fromObjc(value: self, change: change, error: error))
-        }
+    public func observe(on queue: DispatchQueue? = nil,
+                        _ block: @escaping (RealmCollectionChange<LinkingObjects>) -> Void) -> NotificationToken {
+        return rlmResults.addNotificationBlock(wrapObserveBlock(block), receiveOn: queue)
     }
 
     // MARK: Frozen Objects
@@ -331,6 +333,10 @@ public struct LinkingObjects<Element: Object> {
     private init(propertyName: String, handle: RLMLinkingObjectsHandle?) {
         self.propertyName = propertyName
         self.handle = handle
+    }
+    internal init(objc: RLMResults<AnyObject>) {
+        self.propertyName = ""
+        self.handle = RLMLinkingObjectsHandle(linkingObjects: objc)
     }
 
     internal var rlmResults: RLMResults<AnyObject> {
@@ -375,12 +381,11 @@ extension LinkingObjects: RealmCollection {
     }
 
     /// :nodoc:
-    public func _observe(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) ->
-        NotificationToken {
-            let anyCollection = AnyRealmCollection(self)
-            return rlmResults.addNotificationBlock { _, change, error in
-                block(RealmCollectionChange.fromObjc(value: anyCollection, change: change, error: error))
-            }
+    // swiftlint:disable:next identifier_name
+    public func _observe(_ queue: DispatchQueue?,
+                         _ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void)
+        -> NotificationToken {
+            return rlmResults.addNotificationBlock(wrapObserveBlock(block), receiveOn: queue)
     }
 }
 
@@ -388,15 +393,11 @@ extension LinkingObjects: RealmCollection {
 
 extension LinkingObjects: AssistedObjectiveCBridgeable {
     internal static func bridging(from objectiveCValue: Any, with metadata: Any?) -> LinkingObjects {
-        guard let object = objectiveCValue as? RLMObjectBase else { preconditionFailure() }
-        guard let propertyName = metadata as? String else { preconditionFailure() }
-
-        var ret = LinkingObjects(fromType: Element.self, property: propertyName)
-        ret.handle = RLMLinkingObjectsHandle(object: object, property: RLMObjectBaseObjectSchema(object)![propertyName]!)
-        return ret
+        guard let object = objectiveCValue as? RLMResults<Element> else { preconditionFailure() }
+        return LinkingObjects<Element>(objc: object as! RLMResults<AnyObject>)
     }
 
     internal var bridged: (objectiveCValue: Any, metadata: Any?) {
-        return (objectiveCValue: handle!.parent, metadata: handle!.property.name)
+        return (objectiveCValue: handle!.results, metadata: nil)
     }
 }