+++ RealmSwift/Results.swift
@@ -111,6 +111,9 @@ public struct Results<Element: RealmCollectionValue>: Equatable {
     internal init(_ rlmResults: RLMResults<AnyObject>) {
         self.rlmResults = rlmResults
     }
+    internal init(objc rlmResults: RLMResults<AnyObject>) {
+        self.rlmResults = rlmResults
+    }
 
     // MARK: Index Retrieval
 
@@ -299,9 +302,10 @@ public struct Results<Element: RealmCollectionValue>: Equatable {
      not perform a write transaction on the same thread or explicitly call `realm.refresh()`, accessing it will never
      perform blocking work.
 
-     Notifications are delivered via the standard run loop, and so can't be delivered while the run loop is blocked by
-     other activity. When notifications can't be delivered instantly, multiple notifications may be coalesced into a
-     single notification. This can include the notification with the initial collection.
+     If no queue is given, notifications are delivered via the standard run loop, and so can't be delivered while the
+     run loop is blocked by other activity. If a queue is given, notifications are delivered to that queue instead. When
+     notifications can't be delivered instantly, multiple notifications may be coalesced into a single notification.
+     This can include the notification with the initial collection.
 
      For example, the following code performs a write transaction immediately after adding the notification block, so
      there is no opportunity for the initial notification to be delivered first. As a result, the initial notification
@@ -336,13 +340,14 @@ public struct Results<Element: RealmCollectionValue>: Equatable {
 
      - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
 
+     - parameter queue: The serial dispatch queue to receive notification on. If
+                        `nil`, notifications are delivered to the current thread.
      - parameter block: The block to be called whenever a change occurs.
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
-    public func observe(_ block: @escaping (RealmCollectionChange<Results>) -> Void) -> NotificationToken {
-        return rlmResults.addNotificationBlock { _, change, error in
-            block(RealmCollectionChange.fromObjc(value: self, change: change, error: error))
-        }
+    public func observe(on queue: DispatchQueue? = nil,
+                        _ block: @escaping (RealmCollectionChange<Results>) -> Void) -> NotificationToken {
+        return rlmResults.addNotificationBlock(wrapObserveBlock(block), receiveOn: queue)
     }
 
     // MARK: Frozen Objects
@@ -386,12 +391,11 @@ extension Results: RealmCollection {
     public func index(before i: Int) -> Int { return i - 1 }
 
     /// :nodoc:
-    public func _observe(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) ->
-        NotificationToken {
-        let anyCollection = AnyRealmCollection(self)
-        return rlmResults.addNotificationBlock { _, change, error in
-            block(RealmCollectionChange.fromObjc(value: anyCollection, change: change, error: error))
-        }
+    // swiftlint:disable:next identifier_name
+    public func _observe(_ queue: DispatchQueue?,
+                         _ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void)
+        -> NotificationToken {
+            return rlmResults.addNotificationBlock(wrapObserveBlock(block), receiveOn: queue)
     }
 }