+++ CHANGELOG.md
@@ -6,6 +6,8 @@ x.y.z Release notes (yyyy-MM-dd)
   within blocks dispatched to that queue regardless of what thread they happen
   to run on. In addition, change notifications will be delivered to that queue
   rather than the thread's run loop.
+* Add an option to deliver object and collection notifications to a specific
+  queue rather than the current thread.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)